back-end/requirements.txt
alembic==1.4.3
astroid==2.4.2
blinker==1.4
click==7.1.2
colorama==0.4.4
coverage==5.3.1
flake8==3.8.4
Flask==1.1.2
Flask-Cors==3.0.9
Flask-HTTPAuth==4.2.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.4
isort==5.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
Mako==1.1.3
MarkupSafe==1.1.1
mccabe==0.6.1
pycodestyle==2.6.0
pyflakes==2.2.0
PyJWT==2.0.0
pylint==2.6.0
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
six==1.15.0
SQLAlchemy==1.3.22
toml==0.10.2
Werkzeug==1.0.1
wrapt==1.12.1
